+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Browserlist support on build time
 
 ## [1.1.9] - 2024-09-01
 ### Changed
@@ -276,7 +278,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Few issues on dark mode and event listener
 
-[unreleased]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.9...HEAD
+[Unreleased]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.9...HEAD
 [1.1.9]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.8...v1.1.9
 [1.1.8]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.7...v1.1.8
 [1.1.7]: https://github.com/RSickenberg/rsickenberg.me/compare/v1.1.6...v1.1.7